// src/redux/statistics/statistics.types.ts
enum StatisticsActionTypes {
  FETCH_DAILY_ORDERS_STATISTIC_REQUESTED = 'FETCH_DAILY_ORDERS_STATISTIC_REQUESTED',
  FETCH_DAILY_ORDERS_STATISTIC_SUCCESS = 'FETCH_DAILY_ORDERS_STATISTIC_SUCCESS',
  FETCH_DAILY_ORDERS_STATISTIC_FAILED = 'FETCH_DAILY_ORDERS_STATISTIC_FAILED',

  FETCH_MONTHLY_ORDERS_STATISTIC_REQUESTED = 'FETCH_MONTHLY_ORDERS_STATISTIC_REQUESTED',
  FETCH_MONTHLY_ORDERS_STATISTIC_SUCCESS = 'FETCH_MONTHLY_ORDERS_STATISTIC_SUCCESS',
  FETCH_MONTHLY_ORDERS_STATISTIC_FAILED = 'FETCH_MONTHLY_ORDERS_STATISTIC_FAILED',

  FETCH_YEARLY_ORDERS_STATISTIC_REQUESTED = 'FETCH_YEARLY_ORDERS_STATISTIC_REQUESTED',
  FETCH_YEARLY_ORDERS_STATISTIC_SUCCESS = 'FETCH_YEARLY_ORDERS_STATISTIC_SUCCESS',
  FETCH_YEARLY_ORDERS_STATISTIC_FAILED = 'FETCH_YEARLY_ORDERS_STATISTIC_FAILED',

  FETCH_TOP_3_BESTSELLING_PRODUCTS_REQUESTED = 'FETCH_TOP_3_BESTSELLING_PRODUCTS_REQUESTED',
  FETCH_TOP_3_BESTSELLING_PRODUCTS_SUCCESS = 'FETCH_TOP_3_BESTSELLING_PRODUCTS_SUCCESS',
  FETCH_TOP_3_BESTSELLING_PRODUCTS_FAILED = 'FETCH_TOP_3_BESTSELLING_PRODUCTS_FAILED'
}

export default StatisticsActionTypes